feat: route reads, writes and parameter binding through codecs
Wire the codec layer into the JDBC implementation surface:

PgResultSet
- Field carries cached BinaryCodec/TextCodec references (getFieldWithCodec).
- getObject(int) and getObject(int, Class<T>) dispatch to
  decodeBinary/decodeText (and decodeBinaryAs/decodeTextAs for typed
  targets), preserving the legacy connection.getObject(typeName,...)
  fallback for PGobject subclasses registered via addDataType.
- Primitive getters (getByte/Short/Int/Long/Float/Double/getNumeric)
  call codec.decodeAsInt/Long/Float/Double/BigDecimal or the
  decodeTextBytesAsInt/Long byte fast path, avoiding String round-trip.
  Boolean-to-numeric is gated by the new convertBooleanToNumeric
  property and enforced uniformly for binary and text inputs through
  BoolCodec.
- DATE/TIME/TIMETZ/TIMESTAMP/TIMESTAMPTZ honor
  prefersJavaTimeFor* by routing through decodeViaCodec, which
  throws explicitly when no codec is registered.
- The wrong binary→text fall-through in getObject is removed; the
  Array.class branch and the deferred T[].class branch are dropped
  (codec dispatch covers Array.class; typed primitive arrays are out
  of scope per the spec).

PgPreparedStatement / PgCallableStatement
- All parameter binding for SQLData / Struct / UUID flows through a
  single bindViaCodec(parameterIndex, value, pgType) helper that
  resolves the codec from CodecRegistry and dispatches binary or text
  via BaseConnection.binaryTransferSend(oid).
- CompositeCodec gains an encodeAttributesAsText / encodeAttributesAsBinary
  pair that delegates per-attribute encoding to the registered codec,
  replacing Object.toString() based serialization.
- setSQLType / registerOutParameter accept JDBC 4.2 SQLType via
  PGSQLType; the typeMap honored on the CallableStatement getXxx path.
- TypeCoercion centralizes the legacy setObject coercion to int /
  long / float / double / BigDecimal / String / Boolean.

PgConnection
- createStruct() is implemented on top of PgStruct; addDataType
  registrations are routed through JavaTypeRegistry.
- PgArray serves array element extraction (getArray, getArray(map),
  getResultSet) through the codec layer.

Core
- PGConnection / BaseConnection expose getCodecContext, getCodecs,
  registerCodec, resetCodecs and the binary-transfer hooks.
- PGProperty adds convertBooleanToNumeric and the map.pg_type.boolean
  override.
- core.Field carries a cached PgType reference so the codec dispatch
  is O(1) per column.
- QueryExecutor / QueryExecutorImpl pick up the small additions used
  for binary-transfer negotiation.

Tests
- GetObject310Test refreshed for the new dispatch.

+++ pgjdbc/src/main/java/org/postgresql/jdbc/codec/OidCodec.java
@@ -140,7 +140,7 @@ public final class OidCodec implements BinaryCodec, TextCodec {
       return (T) String.valueOf(value);
     }
     if (targetClass == Double.class) {
-      return (T) Double.valueOf(value);
+      return (T) Double.valueOf((double) value);
     }
     if (targetClass == BigDecimal.class) {
       return (T) BigDecimal.valueOf(value);
